feat: Create a new discover page
This commit creates a new discover page and moves the paper list from the home page to the new page. The "Explore Papers" button and the "Discover" link in the header now point to the new page.
